refactor(grpc): decompose grpc_service.rs into submodules — Fixes EVE-101
Split the 4,339-line gRPC service into:
- mod.rs (812 lines): imports, auth/TLS helpers, struct definition,
  WorkerServiceImpl helper methods
- worker_service_impl.rs (3,203 lines): #[tonic::async_trait] impl
  WorkerService trait implementation
- tests.rs (336 lines): all unit tests

The trait impl remains in a single file (Rust requires trait impls in
one block) but is cleanly separated from the struct definition and
helper code. Session-to-Proto duplication already resolved via
schema_session_to_proto() in internal-protocol.

All 28 tests pass. No behavioral changes.
